test(refactor): focus stream test to one case each
stream_test had 3 cases

multi turn is delted since we kinda do that in handoff already split tool
call and basic text into two tests makes it easier with the way that handoff
works to do one test per file
