+++ test/integration/live/handoff_stream_tool_call_test.rb
@@ -4,11 +4,11 @@ require "test_helper"
 require "json"
 require_relative "../../utils/live_test_helper"
 
-class HandoffStreamLiveTest < Test
+class HandoffStreamToolCallLiveTest < Test
   include LiveTestHelper
 
-  SOURCE_TEST_PATH = File.expand_path("stream_test.rb", __dir__)
-  FIXTURE_DIR = File.expand_path("../../fixtures/handoff/stream_test", __dir__)
+  SOURCE_TEST_PATH = File.expand_path("stream_tool_call_test.rb", __dir__)
+  FIXTURE_DIR = File.expand_path("../../fixtures/handoff/stream_tool_call_test", __dir__)
 
   PAIRS = eval(File.read(SOURCE_TEST_PATH).match(/PAIRS = (\[.*?\])\s*\.freeze/m)[1]).freeze
 
@@ -20,10 +20,9 @@ class HandoffStreamLiveTest < Test
     records = load_recorded_outputs
 
     prompt = <<~PROMPT
-      You are receiving recorded final outputs from previous general streaming test runs.
-      The records include text streaming and tool-use streaming results. Read the JSON and answer in one
-      short sentence. Include all of these facts as Arabic numerals where numeric: the number of recorded
-      outputs, the tool-call sum 15 + 27, the multiplication result 42 * 17, and the sum 453 + 434.
+      You are receiving recorded final outputs from previous streaming tool-call test runs.
+      Read the JSON and answer in one short sentence. Include all of these facts as Arabic numerals
+      where numeric: the number of recorded outputs and the tool-call sum 15 + 27.
 
       #{JSON.pretty_generate(records)}
     PROMPT
@@ -34,12 +33,10 @@ class HandoffStreamLiveTest < Test
     text = response.content.select { |block| block.type == "text" }.map(&:text).join(" ").downcase
     assert_includes text, records.length.to_s
     assert_includes text, "42"
-    assert_includes text, "714"
-    assert_includes text, "887"
   end
 
   def self.define_handoff_stream_test_for(provider:, model:)
-    test "handoff_stream__#{provider}_#{model}" do
+    test "handoff_stream_tool_call__#{provider}_#{model}" do
       with_vcr_adapter(provider:, model:) do |adapter|
         run_handoff_stream_for(provider:, model:, adapter:)
       end
@@ -53,7 +50,7 @@ class HandoffStreamLiveTest < Test
   private
 
   def load_recorded_outputs
-    skip "Missing fixture directory at #{FIXTURE_DIR}. Run stream_test live tests first." unless Dir.exist?(FIXTURE_DIR)
+    skip "Missing fixture directory at #{FIXTURE_DIR}. Run stream_tool_call_test live tests first." unless Dir.exist?(FIXTURE_DIR)
 
     Dir.glob(File.join(FIXTURE_DIR, "*.json")).sort.map do |path|
       {
@@ -61,7 +58,7 @@ class HandoffStreamLiveTest < Test
         result: JSON.parse(File.read(path))
       }
     end.tap do |records|
-      skip "No recorded outputs in #{FIXTURE_DIR}. Run stream_test live tests first." if records.empty?
+      skip "No recorded outputs in #{FIXTURE_DIR}. Run stream_tool_call_test live tests first." if records.empty?
     end
   end
 end
